Move readSavedFile() from Nimbus to Storage Class
