+++ src/main/java/Nimbus.java
@@ -12,6 +12,7 @@ public class Nimbus {
     final private static String DATA_FILE_PATH = "./data/data.txt";
 
     static Ui ui;
+    static Storage storage;
 
     public enum Command {
         Remove,
@@ -207,53 +208,21 @@ public class Nimbus {
         }
     }
 
-    // create all the necessary directory and file if the file doesn't exists
-    public static void checkSavedFile(String filePath) {
-        File file = new File(filePath);
-        if (!file.getParentFile().exists()) {
-            file.getParentFile().mkdirs();
-        }
-        if (!file.exists()) {
-            try {
-                file.createNewFile();
-            } catch (IOException e) {
-                System.out.println("Error occur when creating file.");
-            }
-        }
-    }
-
-    public static void readSavedFile(String filePath) {
-        checkSavedFile(filePath);
-        File file = new File(filePath);
-        try {
-            Scanner sc = new Scanner(file);
-            while (sc.hasNext()) {
-                Task task = getTaskFromSavedCommand(sc.nextLine());
-                if (task != null) {
-                    addTask(task, false, false);
-                }
-            }
-        } catch (FileNotFoundException e) {
-            System.out.println("Can't find saved data file");
-        }
-    }
-
-    public static Task getTaskFromSavedCommand(String command) {
-        String[] arr = command.split("\\|");
-        boolean isDone = arr[1].equals("1");
-        return switch (arr[0]) {
-            case "T" -> new Todo(arr[2], isDone);
-            case "D" -> new Deadline(arr[2], isDone, arr[3]);
-            case "E" -> new Event(arr[2], isDone, arr[3], arr[4]);
-            default -> null;
-        };
+    private static void readSavedFile() {
+        ArrayList<Task> arr = storage.readSavedFile();
+        tasks.addAll(arr);
     }
 
     public static void main(String[] args) {
         Nimbus.ui = new Ui(name);
-        ui.showWelcomeMessage();
+        Nimbus.storage = new Storage(DATA_FILE_PATH);
 
-        readSavedFile(DATA_FILE_PATH);
+        run();
+    }
+
+    public static void run() {
+        readSavedFile();
+        ui.showWelcomeMessage();
 
         Scanner scanner = new Scanner(System.in);
         String line;
@@ -268,5 +237,6 @@ public class Nimbus {
         }
 
         ui.showGoodbyeMessage();
+
     }
 }
